server/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
pydantic[email]
mangum
